requirements.txt
wheel
carpi-daemoncommons
carpi-redisdatabus
gps3
